Update patch set 3
Patch Set 3:

The scenario came up with XACML PDp and guard policies. Some of the values are Integer, but GSON was converting them to doubles. This broke the XACML because we are expecting integer. It will break the DCAE Monitoring policies or any policy that has properties that are expected to be Integer vs Double.

Patch-set: 3