Update patch set 2
Patch Set 2:

> 185 82.7% let's see the improvements!
 > 
 > run-sonar

I don't expect a lot as most of these were in the Test classes, which only show up in SonarLint, not in plain sonar.  However, one of the code fixes (for the "closeCondition") was actually a bug fix, based on javadocs.  Of course, now it complains that the code ignores the return value from await(), so it may be a wash.

Patch-set: 2