Update patch set 5
Patch Set 5:

> > Patch Set 5:
 > >
 > > > integrity-audit: is this something that only drools and
 > >  > drools-applications care about?
 > >
 > >  > integrity-audit: is this something that only drools and
 > >  > drools-applications care about?
 > >
 > > The features that use this functionality are disabled in the
 > PDP-D side.
 > 
 > Ok, just wondering if these should be moved from policy/common to
 > drools-applications?? Not sure if the other components would need
 > these??

I think the reason it is there is because policy/engine uses this in some capability.

 > > Patch Set 5:
 > >
 > > > integrity-audit: is this something that only drools and
 > >  > drools-applications care about?
 > >
 > >  > integrity-audit: is this something that only drools and
 > >  > drools-applications care about?
 > >
 > > The features that use this functionality are disabled in the
 > PDP-D side.
 > 
 > Ok, just wondering if these should be moved from policy/common to
 > drools-applications?? Not sure if the other components would need
 > these??

The issue is that policy/engine uses it, so will have to keep it here until policy/engine deprecates.
I don't think other components will use it, even in default onap configuration, I don't see a good reason to enable it, but giving the choice to clients to use it in their customized deployments.

Patch-set: 5