Fix managed noop topic source
NOOP topic sources configured via TopicParameterGroup are left
unmanaged.  Removed a new override method, as it did not pass
the "managed" flag down, while the method that already existed
in the superclass already works correctly.

Also restored "null" lists to TopicParameterGroup so-as not
to break the behavior of its validate() method.

Change-Id: Id392a3263c3bd1c8f278461715b441719817c47c
Issue-ID: POLICY-1953
Signed-off-by: Jim Hahn <jrh3@att.com>