Update patch set 1
Patch Set 1:

I am not sure about these changes, I was only expecting changes in CryptoUtils, basically to have a holder for the key that would be given to it at instantiation.    There should be no changes in "common-logging" but you would have some simple changes in PropertyUtils in policy/drools-pdp to do the expansion.    With regards to the PropertyCoder, need to talk to Jim, I don't remember the details of what we agreed, and not sure if this is related, in any case, would assume decrypt operations would be done by calls to the Crypto.. object.

Patch-set: 1