Update patch set 4
Patch Set 4: Code-Review-1

A few general comments:

1.  No changes should be necessary under "common-logging" flexlogger package.      I doubt there are any uses in any repos for encrypted data that relates to flexlogger.   FlexLogger is not recommended in any case.   This file seems like a plain copy of the one in drools-pdp repo.   Therefore let's leave this functionality untouched here, so all that code I think should go away.   Eventually should get rid of that class.

2.  I believe we should not use environment variables in a generic library, commons is for generic libraries, as this goes beyond imposing constraints in java code, it imposes constraints to the installer, and the OS where this library will be run.

3.  I started to put together some code in an in-progress review, see:  https://gerrit.onap.org/r/#/c/policy/drools-pdp/+/97229/.  I
ll be adding you and Jim while in progress to discuss.  Note that there the code is very simple and will replace all the code under flexlogger package for the drools-pdp.   I will add next getting the secret key in the drools pdp, so you have it as a reference.   In the future, we may want to move all that to commons but is OK for now.   We can discuss engine in a later iteration.

Patch-set: 4
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1