Update patch set 1
Patch Set 1:

This is not where we define plugin versions - the versions are defined in oparent firstly, and secondarily in policy/parent. As well as which ones are used, their configuration. BEFORE you go and override them in each repo.

Please be aware this repo is being upgraded to JDK 11.

Patch-set: 1