Update patch set 4
Patch Set 4:

> I'll defer to Jorge on this.  I'm not really comfortable adding
 > this as additional methods on HttpClient, but maybe it's OK,
 > provided we fix it next release.

I think Jim's proposal to move the functionality out of JettyJerseyServer is best under the following assumption:

"We won't have cases (either no technically possible, or design wise not desirable) where a (jaxrs) Jersey Servlet and a (resource) DefaultServlet will coexist under the same umbrella HTTP Server.   So in other words are mutually exclusive, and from that it is a better approach to separate it into different entities, classes.".  

Would this be a fair way to put it Jim?

Patch-set: 4