rename jpa persistence file
After inspecting the fat jar generated by brmsgw, noticed that
the persistence.xml included from here it was overriden by another
one, therefore the persistence unit was not available.   This
may solve the issue for the brmsgw, as with a far jar build, there
are the risks of collisions such as this of files with the same
name.

Issue-ID: POLICY-2471
Signed-off-by: jhh <jorge.hernandez-herrero@att.com>
Change-Id: I91cf2b6d3d2f5086d77795e08800ca96d984c214