Update patch set 1
Patch Set 1:

This is for the brmsgw not coming up sane, there are some database issues.   The legacy healthcheck will flag it when querying the PDP (as it is dependent).    The legacy PDP healthchek is disabled right now in the healthcheck, that's the reason it doesn't show as an issue in the reports.

Patch-set: 1