Update patch set 3
Patch Set 3:

Hi Jim,

The cost of the method is we could not call Java API freely in JavaScript.

In my testing, with below script
https://github.com/mozilla/rhino/blob/master/examples/liveConnect.js
It throws exception like this:
TypeError: Cannot find function append in object

Another option I have investigated was to customize JSON.stringify, so that it may be possible to integrate jackson to do the conversion. But Seems Rhino does not support this.

Patch-set: 3