Add bean validator based on javax.validation
Added a new bean validator that uses the javax.validation API.  The
original validator and its associated annotations will be left in the
source code until the other repos have been modified to use the new
validator.
Uses hierarchical format for messages, as that facilitated adding nested
messages to the results.
Modified the old validator to support both annotation types.
Note: in javax.validation, the NotBlank annotation has an implicit
NotNull check, while the validator in policy-common does not.  Added a
NullNotBlank annotation that can be used instead where the old behavior
is needed.
Note: in some cases, it appeared that class-level support of NotNull may
not be available with javax.validation.

Issue-ID: POLICY-2059-javax-valid
Change-Id: I2111e985d9725313fb3818ecc7fa62732aa040ee
Signed-off-by: Jim Hahn <jrh3@att.com>