Wait after fetch exception on topic
When dmaap is inaccessible for some reason, the topic source frequently
enters a fast fail loop, rapidly filling up the log.  Modified the code
to wait the configured fetchTimeout when this occurs.

With any luck, this will also fix the sporadic kubernetes crash-fail
loops sometimes seen with the xacml-pdp pod.

Issue-ID: POLICY-3457
Change-Id: Icf45237a6eeec35ea8a938f1630b0477f2084c56
Signed-off-by: Jim Hahn <jrh3@att.com>