Integrating prometheus with Policy components
This is just a draft review, but I would like to get feedback from
everyone about this. So, making this review ACTIVE.

With this change, all PF components can add servlets by providing the
servletPath and class.
For example to add prometheus metrics servlet in policy-api:
if (params.isExportMetrics()) {
  restServer = new RestServer(params, AafApiFilter.class,
      Map.of("/metrics", MetricsServlet.class),
      ApiRestController.class);
}

In addition, we can later go to individual components and add more
metrics, say for e.g.,
deployedPoliciesCounter/undeployedPoliciesCounter etc on PAP,
executedEvents/failedEvents etc on PDP and so on.
This will look something like below, for e.g. in policy-pap component:
 io.prometheus.client.Counter counter = Counter.build()
  .name("policies_deployed_total")
  .help("Number of policies deployed.").register()
Whenever a policy is deployed, just call
 counter.inc()

Usage of code like above in the individual component will expose
such data as well as part of the exposed metrics.

Change-Id: Id667f27b15c012398421ba657b5324cc1d82cf1f
Issue-ID: POLICY-3524
Signed-off-by: a.sreekumar <ajith.sreekumar@bell.ca>
Signed-off-by: jhh <jorge.hernandez-herrero@att.com>
Signed-off-by: a.sreekumar <ajith.sreekumar@bell.ca>