Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (4 comments)
> 
> Thanks Ajith, this is great.   Since the Prometheus servlet for measurements is just a plain servlet. 
>   I am thinking may be better to keep the http server endpoint agnostic of which type of servlet it is.   I thought it was supporting already out of the box but may be not.  It currently supports servlet filters (see.  JettyServletServer.addFilter - that AuthorizationFilter uses) and jax-rs endpoints.    I think adding generic servlet support if already not there would be great, this would be similar to the JettyServletServer.addFilter generic method.    This would be out of the box support for using the prometheus MetricServlet.

Agree with your comments Jorge.
But I am not sure if we can make it generic to that level. Otherwise we will have to pass the Servletpath (/metrics in this case) and ServletClass (io.prometheus.client.exporter.MetricsServlet) from each component. If this should be generic enough to support multiple cases, then a list of <Servletpath, ServletClass> pair will have to be provided from each component. Should we go this way! It might be a little overkill. If atleast we go ahead with the new Servlet as one for Metrics, then we can go with "/metrics" and use whichever provider class of the metricsServlet the respective component passes.

Another option will be to go to each component and  write a jaxrs endpoint for collecting metrics. This doesn't seem to the preferred option though. Preferred option is to use MetricsServlet AFAIK.

Patch-set: 2
Reviewer: Gerrit User 4903 <4903@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>