Update patch set 7
Patch Set 7:

I just uploaded a patch to this CI, Ajith, take a look at patch 7.   Best way to look at it is to look the junits that introduces several configuration possibilities.  I think these changes give a lot of flexibility to the individual component to set it up simply as they wish.  It provides generic support for non-JAX-RS servlets which is reused by the static server flavor.  It needs a little work on the RestServer which follows the original properties configuration that we use in drools, but I think if you want to keep it simple, just a boolean to enable the prometheus server will simply do it.   In any case, feel free to modify, or discard the code as you want.

Patch-set: 7
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>