Update patch set 8
Patch Set 8:

(1 comment)

> Patch Set 7:
> 
> I just uploaded a patch to this CI, Ajith, take a look at patch 7.   Best way to look at it is to look the junits that introduces several configuration possibilities.  I think these changes give a lot of flexibility to the individual component to set it up simply as they wish.  It provides generic support for non-JAX-RS servlets which is reused by the static server flavor.  It needs a little work on the RestServer which follows the original properties configuration that we use in drools, but I think if you want to keep it simple, just a boolean to enable the prometheus server will simply do it.   In any case, feel free to modify, or discard the code as you want.

I see that you cleaned it up, especially adding the servlet only once. I was planning to do such things once the main changes are finalized. Anyway that is cool, Thanks.
I think we don't need an extra way to initialize prometheus metrics specifically in policy-common when we support the initializing Servlets in a generic way with servletPath. So, removed such changes in patchset 8.
But my main question is, as per patchset6 I did, the way to add any servlet was by specifying them in the constructor from any component that uses it.
For example to add prometheus metrics servlet in policy-api:
if (params.isExportMetrics()) {
  restServer = new RestServer(params, AafApiFilter.class,
      Map.of("/metrics", MetricsServlet.class),
      ApiRestController.class);
}
where exportMetrics will be a boolean in the component configuration file.

With the latest changes of moving the servletPath and servletClass to properties,, how are we initializing new Servlets? We should add another entry for these details in restServerParameters of components. Otherwise, again we have to go as I mentioned above as per patchset6 (in which case some of the changes related to properties after patch6 may be removed)

Patch-set: 8
Reviewer: Gerrit User 4903 <4903@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>