Update patch set 8
Patch Set 8:

Oh, Ajith, I think you deleted some of my changes (patch 7) with your last patch (patch 8), not sure if purposefully, in any case patch 7 is that's what I described with my comments, the idea was to model after aaf (which adds a filter blindly with a boolean configuration flag), and allow for a generic way too that allows provision of any servlet (which can be used for the MetricsServlet), and allow for standalone servers with nothing else but the metrics server.   My thinking is that would cover most of the scenarios that one would want to cover.  I think this aligns too with the endpoints http server design.  In theory, JettyJerseyServlet should be touched minimally as it is concerned with integrating its own jersey servlet, and only used for jax-rs (rest) stuff and swagger (which is for jax-rs too). Generic servlet work should go in the parent class JettyServletServer and that is where  filters, auth, and generic servlet functionality is dealt with, just some background info.   
In any case, let me know your thoughts and we can discuss further whatever we are not aligned.

Patch-set: 8
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>