Fix test failure in staging job
The "keytool" program does not exist in the Jenkins staging job
environment, so the test case that checks for a keytool error message
fails because an unexpected error message is received.

By changing the test case to expect "keytool" rather than "keytool
exited with", then the test will pass when the error message "Cannot
run program "keytool": error=2, No such file or directory" is returned.

Issue-ID: POLICY-3587
Change-Id: I3a6b77fb571cfbf170f5f16bf93eece2d78a7bbb
Signed-off-by: liamfallon <liam.fallon@est.tech>